add an extra factory to solve data-sharing problem
Problem:
Since $http response data cannot be resolved directly in a FACTORY and it is only possible to return a promise object, there is a problem about sharing data "quizQuestions".  More specifically, some properties of quizQuestions are changed in the quiz controller. These changes cannot be obtained in the result controller.

Idea:
Use an extra factory containing all necessary data in an optimized format to share data.
---------------------------------------------------------
It didn't work !!!